refactor(site): page profile fund item
