+++ packages/site/src/pages/Profile.js
@@ -7,7 +7,6 @@ import { useParams } from "react-router";
 import serverApi from "../services/serverApi";
 import { newErrorToast } from "../store/reducers/toastSlice";
 import { useDispatch } from "react-redux";
-import FundsList from "../components/Profile/FundsList";
 import ProfileDataList from "../components/Profile/ProfileDataList";
 
 const ContentWrapper = styled.div`
@@ -50,8 +49,6 @@ export default function Profile() {
       <Container>
         <ContentWrapper>
           <ProfileDataList tab={tab} network={network} address={address} />
-
-          {tab === "funds" && <FundsList network={network} address={address} />}
         </ContentWrapper>
       </Container>
     </div>
